requirements.txt
africastalking==1.2.8
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
defusedxml==0.8.0rc2
Django==4.2.16
django-allauth==64.2.1
djangorestframework==3.15.2
ecdsa==0.19.0
idna==3.10
josepy==1.14.0
mozilla-django-oidc==4.0.1
oauthlib==3.2.2
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schema==0.7.7
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
